refactor: ♻️ (mobile) refactor theme feature
