+++ apps/mobile/src/features/theme/index.ts
@@ -1,3 +1,3 @@
 export { CustomThemeProvider, ThemeProvider } from "./config/provider";
 export { useTheme, themeAtom } from "./models/theme";
-export type { ColorTheme, ThemeStorage } from "./types/theme.type";
+export type { ColorTheme, ThemeStorage } from "./types/theme";
